net/mlx5: support more than 4M conntrack objects
In order to support more than 4M conntrack objects, the apportionment
of the CT index is changed in this commit.

The maximal CT number supported is changed from 4M to 32M, by using
another 3 more bits. In the meanwhile, the maximal ports supported
is reduced to 32 by removing 3 bits. The 2 MSB is still used as the
indirect action type.

The CT pool management is also extended by using u32 instead of u16
as the type of pool index. More than 64K pools can be supported and
there is still 64 CT objects in one pool due to the bulk number.

Some overflow mechanism is used for the last object in the last pool,
and all the least 25 bits with 0 is used as the last one.

+++ drivers/net/mlx5/mlx5.h
@@ -1134,7 +1134,7 @@ struct mlx5_aso_ct_action {
 
 /* ASO connection tracking software pool definition. */
 struct mlx5_aso_ct_pool {
-	uint16_t index; /* Pool index in pools array. */
+	uint32_t index; /* Pool index in pools array. */
 	struct mlx5_devx_obj *devx_obj;
 	/* The first devx object in the bulk, used for freeing (not yet). */
 	struct mlx5_aso_ct_action actions[MLX5_ASO_CT_ACTIONS_PER_POOL];
@@ -1148,8 +1148,8 @@ LIST_HEAD(aso_ct_list, mlx5_aso_ct_action);
 /* Pools management structure for ASO connection tracking pools. */
 struct mlx5_aso_ct_pools_mng {
 	struct mlx5_aso_ct_pool **pools;
-	uint16_t n; /* Total number of pools. */
-	uint16_t next; /* Number of pools in use, index of next free pool. */
+	uint32_t n; /* Total number of pools. */
+	uint32_t next; /* Number of pools in use, index of next free pool. */
 	rte_spinlock_t ct_sl; /* The ASO CT free list lock. */
 	rte_rwlock_t resize_rwl; /* The ASO CT pool resize lock. */
 	struct aso_ct_list free_cts; /* Free ASO CT objects list. */
